Move UCR adapter class
